goblin/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.39.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
celery==5.4.0
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
daphne==4.1.2
dataclasses-json==0.6.5
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.7
django-celery-beat==2.7.0
djangorestframework==3.15.2
dnspython==2.6.1
docker==7.1.0
elevenlabs==1.9.0
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
langchain==0.2.11
langchain-community==0.2.9
langchain-core==0.2.24
langchain-nomic==0.1.2
langchain-ollama==0.1.0
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langgraph==0.1.15
langserve==0.2.2
langsmith==0.1.93
libclang==18.1.1
llama-index==0.10.33
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.34
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.3
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.20
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
loguru==0.7.2
lupa==2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
modal==0.64.64
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
nomic==3.1.1
numpy==1.26.4
ollama==0.3.0
openai==1.32.0
openai-whisper==20231117
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pypdf==4.2.0
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.1.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.29
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
sympy==1.13.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
tree-sitter==0.23.0
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.3
uvloop==0.19.0
vosk==0.3.44
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zep-cloud==2.1.0
